+++ src/js/script.js
@@ -1,4 +1,4 @@
 import App from './App';
 
-document.addEventListener('DOMContentLoaded', App.init());
+document.addEventListener('DOMContentLoaded', App.init);
